Rebase new UI with Game object.
