+++ server.py
@@ -1,10 +1,12 @@
 import json
+import uuid
 
-from flask import Flask, jsonify
+from flask import Flask, jsonify, abort
 from flask_cors import CORS
 
 from catanatron.game import Game
 from catanatron.models.map import Water, Port
+from catanatron.models.player import RandomPlayer, Color
 
 
 app = Flask(__name__)
@@ -49,7 +51,22 @@ def serialize_game(game):
     return {"tiles": tiles, "nodes": nodes, "edges": edges}
 
 
-@app.route("/board")
-def board():
-    game = Game()  # Make new in-memory Game, in the future we read board#i
+games = {}
+
+
+@app.route("/games", methods=["POST"])
+def create_game():
+    game = Game(players=[RandomPlayer(Color.RED), RandomPlayer(Color.BLUE)])
+    game_id = uuid.uuid4()
+    games[str(game_id)] = game
+    return jsonify({"game_id": game_id})
+
+
+@app.route("/games/<string:game_id>", methods=["GET"])
+def get_game(game_id):
+    print(games.keys())
+    if game_id not in games:
+        abort(404, description="Resource not found")
+
+    game = games[game_id]
     return jsonify(serialize_game(game))
