+++ ui/src/pages/HomePage.scss
@@ -1,0 +1,11 @@
+.HomePage {
+  height: 60%;
+  max-height: 100%;
+
+  display: flex;
+  flex-direction: column;
+  max-width: 300px;
+  justify-content: center;
+  align-items: center;
+  margin: 0 auto;
+}
